Revert "GOATS-1126: Cache proposals and instruments per user. (#525)"
This reverts commit 967ad696dbdfba73b5efe392c15002c15a71967c.
