+++ src/goats_tom/facilities/__init__.py
@@ -1,10 +1,6 @@
 from .gemini import GEMObservationForm, GOATSGEMFacility
-from .overrides import BLANCOFacility, LCOFacility, SOARFacility
 
 __all__ = [
     "GEMObservationForm",
     "GOATSGEMFacility",
-    "LCOFacility",
-    "SOARFacility",
-    "BLANCOFacility",
 ]
